+++ R/bm_compose.R
@@ -41,6 +41,5 @@ bm_compose <- function(bml, pua_combining = character(0), ...) {
             names(bml_new)[i_new] <- names(bml)[i]
         }
     }
-    class(bml_new) <- c("bm_list", class(bml_new))
-    bml_new
+    as_bm_list.list(bml_new)
 }