+++ R/combine.R
@@ -1,21 +1,28 @@
 #' Combine bitmap objects
 #'
-#' `c()` combines bitmap objects into `bm_list()` or `bm_font()` objects.
+#' `c()` combines bitmap objects into [bm_list()] or [bm_font()] objects.
 #'  In particular when using it to combine fonts the later fonts
 #'  "update" the glyphs in the earlier fonts.
 #'
 #' The various bitmap objects are "reduced" in the following ways:
 #' \tabular{lll}{
 #' First \tab Second \tab Result\cr
-#' `bm_bitmap()` \tab `bm_bitmap()` \tab `bm_list()`\cr
-#' `bm_bitmap()` \tab `bm_list()` \tab `bm_list()`\cr
-#' `bm_bitmap()` \tab `bm_font()` \tab `bm_font()`\cr
-#' `bm_font()` \tab `bm_bitmap()` \tab `bm_font()`\cr
-#' `bm_font()` \tab `bm_list()` \tab `bm_font()`\cr
-#' `bm_font()` \tab `bm_font()` \tab `bm_font()`\cr
-#' `bm_list()` \tab `bm_bitmap()` \tab `bm_list()`\cr
-#' `bm_list()` \tab `bm_list()` \tab `bm_list()`\cr
-#' `bm_list()` \tab `bm_font()` \tab `bm_font()`\cr
+#' [bm_bitmap()] \tab [bm_bitmap()] \tab [bm_list()]\cr
+#' [bm_bitmap()] \tab [bm_font()] \tab [bm_font()]\cr
+#' [bm_bitmap()] \tab [bm_list()] \tab [bm_list()]\cr
+#' [bm_bitmap()] \tab [bm_pixmap()] \tab [bm_list()]\cr
+#' [bm_pixmap()] \tab [bm_bitmap()] \tab [bm_list()]\cr
+#' [bm_pixmap()] \tab [bm_font()] \tab `ERROR`\cr
+#' [bm_pixmap()] \tab [bm_list()] \tab [bm_list()]\cr
+#' [bm_pixmap()] \tab [bm_pixmap()] \tab [bm_list()]\cr
+#' [bm_font()] \tab [bm_bitmap()] \tab [bm_font()]\cr
+#' [bm_font()] \tab [bm_font()] \tab [bm_font()]\cr
+#' [bm_font()] \tab [bm_list()] \tab [bm_font()]\cr
+#' [bm_font()] \tab [bm_pixmap()] \tab `ERROR`\cr
+#' [bm_list()] \tab [bm_bitmap()] \tab [bm_list()]\cr
+#' [bm_list()] \tab [bm_font()] \tab [bm_font()]\cr
+#' [bm_list()] \tab [bm_list()] \tab [bm_list()]\cr
+#' [bm_list()] \tab [bm_pixmap()] \tab [bm_list()]\cr
 #' }
 #' When combining with a `bm_font()` object if any `bm_bitmap()` objects
 #' share the same name we only keep the last one.
@@ -41,8 +48,8 @@ c.bm_bitmap <- function(...) {
         r <- bm_list(...)
     } else {
         second <- l[[2L]]
-        stopifnot(is_bm_bitmap(second) || is_bm_list(second))
-        if (is_bm_bitmap(second)) {
+        stopifnot(is_bm_matrix(second) || is_bm_list(second))
+        if (is_bm_matrix(second)) {
             r <- as_bm_list.list(l[1:2])
         } else if (is_bm_font(second)) {
             r <- c.bm_font(as_bm_font(as_bm_list.list(l[1L])), second)
@@ -95,8 +102,8 @@ c.bm_list <- function(...) {
         r <- l[[1]]
     } else {
         second <- l[[2L]]
-        stopifnot(is_bm_bitmap(second) || is_bm_list(second))
-        if (is_bm_bitmap(second)) {
+        stopifnot(is_bm_matrix(second) || is_bm_list(second))
+        if (is_bm_matrix(second)) {
             r <- c.bm_list(l[[1]], as_bm_list.list(l[2L]))
         } else if (is_bm_font(second)) {
             r <- c(as.list(l[[1]]), as.list(l[[2]]))
@@ -104,6 +111,30 @@ c.bm_list <- function(...) {
             r <- as_bm_font.list(r)
         } else {
             r <- as_bm_list.list(c(as.list(l[[1]]), as.list(l[[2]])))
+        }
+    }
+    if (length(l) > 2L) {
+        do.call(c, c(list(r), l[c(-1L, -2L)]))
+    } else {
+        r
+    }
+}
+
+#' @rdname combine
+#' @export
+c.bm_pixmap <- function(...) {
+    l <- list(...)
+    stopifnot(length(l) > 0L)
+    if (length(l) == 1L) {
+        r <- bm_list(...)
+    } else {
+        second <- l[[2L]]
+        stopifnot(is_bm_matrix(second) || is_bm_list(second),
+                  !is_bm_font(second))
+        if (is_bm_matrix(second)) {
+            r <- as_bm_list.list(l[1:2])
+        } else {
+            r <- c.bm_list(as_bm_list.list(l[1L]), second)
         }
     }
     if (length(l) > 2L) {